Some Grammar Stuffs
I just added a few things to flow better and edits to the comma usage. Some of the changes were to how the sentences were worded, but most of the edits were commas. Since there was variance in the use of the Oxford commas (e.g. Adam, Monique, and Erina (after monique)), I added it to all of the lists. Because our other papers are in APA and APA requires the comma, I chose that option. Other commas I added or took out were because they grammatically should have or should not have gone there.